TYPESCRIPT WORKS
So I figured out how to make it work. The problem was with how the ShardingManager was spawning the shards. It needed a register ts-node thingy to make it work. It works now. Testing will probably work too, but i'm tired of it, so i'm not going to do testing yet.
Also it's hard to test Discord stuff without recreating a whole guild, members, etc, inside the code, just easier to create a testing server and do stuff there.