UPDATE
SO there were several issues here and there, especially with how tests work.
It took a while but I finally got a "right" test config. The problem is we will have to first build the ts project and then run the tests as normal JS tests.
The reason is, I got f***king tired of trying doing TS tests....