Extreme Anger
So, because it doesn't say it anywhere, I LOOKED, BY GOD I LOOKED.
That even the localized names have to be LOWERCASE and SANITIZED so they don't include whitespace.
I had to run the program and test it because JEST didn't show that in the console, no it just said "INVALID STRING FORMAT", okay dipshit, thanks, HOW ABOUT SHOWING ME WHATS WRONG IN THE FORMAT. I looked at regex and figured it out.

SO THANK YOU FOR WASTING LITERAL 2 DAYS ON THIS OMG.