UPDPATE - TESTS
I had to move tests outside of src, because no need to bundle them with dist or dev.
I COULD just make a config to exclude it with tsconfig, but for distribution I want to use tsup, and it doesn't handle that....
So instead I move it out and yeah. I PROBABLY could make it a TS and have it build but... eh